fix check for root data resource in URI
For a vector v of length n, element v[n] is not guaranteed to be NULL
as vector_grow does not clear newly allocated memory.  Get the length
of the vector instead of checking for a NULL element.  This (a) fixes a
segfault when the unitialized pointer value is non-null and (b) correctly
detects the root node so that a POST to /restconf/data creates a new
top-level node.

Signed-off-by: Eric Kinzie <ekinzie@labn.net>